move package tetherjson into tether
