+++ src/go/tether/tetherjson/tetherCmd.go
